+++ client/classes/MediaDbConnection.ts
@@ -1,12 +1,15 @@
 import {
   Movie,
   MovieCollection,
-  Show
+  Show,
+  SemVer
 } from 'mediadb-shared';
 
 import {
   URL
 } from 'url';
+
+import { MissingEndpointError } from './errors/MissingEndpointError';
 
 import { type SearchQuery } from './SearchQuery';
 
@@ -22,18 +25,36 @@ export class MediaDbConnection {
 
   async getMovie(id: string): Promise<Movie> {
     const res = await fetch(`${this.endpoint}/movies/${id}`);
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     const data = await res.json();
     return new Movie(data);
   }
 
   async getMovieCollection(id: string): Promise<MovieCollection> {
     const res = await fetch(`${this.endpoint}/movie-collections/${id}`);
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     const data = await res.json();
     return new MovieCollection(data);
   }
 
   async getShow(id: string): Promise<Show> {
     const res = await fetch(`${this.endpoint}/shows/${id}`, );
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     const data = await res.json();
     return new Show(data);
   }
@@ -44,9 +65,14 @@ export class MediaDbConnection {
     Object.entries(query)
       .forEach(prop => url.searchParams
         .append(prop[0], <string>prop[1]));
-    console.log(url.toString());
     
     const res = await fetch(url);
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     const data = <Partial<Movie>[]>(await res.json());
     return data.map(movie => new Movie(movie));
   }
@@ -57,9 +83,14 @@ export class MediaDbConnection {
     Object.entries(query)
       .forEach(prop => url.searchParams
         .append(prop[0], <string>prop[1]));
-    console.log(url.toString());
     
     const res = await fetch(url);
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     const data = <Partial<MovieCollection>[]>(await res.json());
     return data.map(movie => new MovieCollection(movie));
   }
@@ -70,9 +101,14 @@ export class MediaDbConnection {
     Object.entries(query)
       .forEach(prop => url.searchParams
         .append(prop[0], <string>prop[1]));
-    console.log(url.toString());
     
     const res = await fetch(url);
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     const data = <Partial<Show>[]>(await res.json());
     return data.map(movie => new Show(movie));
   }
@@ -85,6 +121,12 @@ export class MediaDbConnection {
         'Content-Type': 'application/json'
       }
     });
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     return (await res.json()).new_id;
   }
   
@@ -96,6 +138,12 @@ export class MediaDbConnection {
         'Content-Type': 'application/json'
       }
     });
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     return (await res.json()).new_id;
   }
   
@@ -107,6 +155,12 @@ export class MediaDbConnection {
         'Content-Type': 'application/json'
       }
     });
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     return (await res.json()).new_id;
   }
 
@@ -118,6 +172,12 @@ export class MediaDbConnection {
         'Content-Type': 'application/json'
       }
     });
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     return (await res.json()).new_id;
   }
 
@@ -129,6 +189,51 @@ export class MediaDbConnection {
         'Content-Type': 'application/json'
       }
     });
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
     return (await res.json()).new_id;
+  }
+
+  async deleteMovie(id: string): Promise<boolean> {
+    const res = await fetch(`${this.endpoint}/movies/${id}`, {
+      method: 'DELETE'
+    });
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
+    return (await res.json()).is_successful;
+  }
+
+  async deleteMovieCollection(id: string): Promise<boolean> {
+    const res = await fetch(`${this.endpoint}/movie-collections/${id}`, {
+      method: 'DELETE'
+    });
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
+    return (await res.json()).is_successful;
+  }
+
+  async deleteShow(id: string): Promise<boolean> {
+    const res = await fetch(`${this.endpoint}/shows/${id}`, {
+      method: 'DELETE'
+    });
+
+    if (res.status === 404) {
+      throw new MissingEndpointError(
+        res.headers.get('X-API-Version') ?? 'v0.0.0');
+    }
+
+    return (await res.json()).is_successful;
   }
 }