+++ shared/index.ts
@@ -3,6 +3,7 @@ export * from './classes/Media';
 export * from './classes/Movie';
 export * from './classes/MovieCollection';
 export * from './classes/Season';
+export * from './classes/SemVer';
 export * from './classes/Show';
 export * from './classes/Source';
 export * from './classes/Translation';