+++ CHANGELOG.md
@@ -11,6 +11,13 @@ project adheres to
 
 //
 
+## 0.2.0 - 2025-03-26
+
+### Added
+
+- Version checking between client and server
+- Delete functionalities for movies, shows and movie collections
+
 ## 0.1.0 - 2025-03-23
 
 ### Added